@@ -21,34 +21,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
+    "@angular/animations": "^7.1.1",
     "@angular/cdk": "^7.1.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/pwa": "^0.9.0-rc.1",
-    "@angular/router": "^7.0.0",
-    "@angular/service-worker": "7.0.0",
+    "@angular/common": "^7.1.1",
+    "@angular/compiler": "^7.1.1",
+    "@angular/core": "^7.1.1",
+    "@angular/forms": "^7.1.1",
+    "@angular/http": "^7.1.1",
+    "@angular/platform-browser": "^7.1.1",
+    "@angular/platform-browser-dynamic": "^7.1.1",
+    "@angular/pwa": "^0.11.0",
+    "@angular/router": "^7.1.1",
+    "@angular/service-worker": "7.1.1",
     "@fortawesome/fontawesome": "^1.1.8",
     "@fortawesome/fontawesome-free-regular": "^5.0.13",
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
-    "@ng-select/ng-select": "^2.3.6",
+    "@ng-select/ng-select": "^2.13.0",
     "apollo-angular": "^1.3.0",
     "apollo-angular-link-http": "^1.2.0",
-    "apollo-cache-inmemory": "^1.2.7",
-    "apollo-client": "^2.4.1",
-    "apollo-link-ws": "^1.0.8",
+    "apollo-cache-inmemory": "^1.3.11",
+    "apollo-client": "^2.4.7",
+    "apollo-link-ws": "^1.0.10",
     "core-js": "^2.5.4",
-    "file-saver": "^1.3.8",
-    "getstream": "^3.19.0",
+    "file-saver": "^2.0.0",
+    "getstream": "^3.23.1",
     "graphql": "^14.0.2",
     "graphql-tag": "^2.9.2",
-    "isotope-layout": "^3.0.6",
-    "jquery": "^3.3.1",
     "npm": "^6.4.1",
     "rxjs": "^6.2.2",
     "subscriptions-transport-ws": "^0.9.14",
@@ -56,29 +54,29 @@
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.9.0-rc.1",
-    "@angular/cli": "^7.0.2",
-    "@angular/compiler-cli": "^7.0.0",
-    "@angular/language-service": "^7.0.0",
-    "@compodoc/compodoc": "^1.1.6",
-    "@types/file-saver": "^1.3.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@angular-devkit/build-angular": "^0.11.0",
+    "@angular/cli": "^7.1.0",
+    "@angular/compiler-cli": "^7.1.1",
+    "@angular/language-service": "^7.1.1",
+    "@compodoc/compodoc": "^1.1.7",
+    "@types/file-saver": "^2.0.0",
+    "@types/jasmine": "~3.3.0",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~10.12.11",
     "@types/uuid": "^3.4.3",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "^3.1.3",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "^5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^3.1.3",
-    "webpack-bundle-analyzer": "^2.9.2"
+    "ts-node": "~7.0.1",
+    "tslint": "~5.11.0",
+    "typescript": "^3.1.6",
+    "webpack-bundle-analyzer": "^3.0.3"
   },
   "prettier": {
     "singleQuote": true,
