@@ -1,14 +1,15 @@
-import { AfterViewInit, ChangeDetectorRef, Component, Input, NgModuleRef, OnInit } from '@angular/core';
+import { AfterViewInit, ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { ERM, Product, ERM_TOKEN } from '~models';
-import { ProductService } from '~entity-services';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
+import { ProductService } from '~entity-services';
+import { ERM, Product } from '~models';
+import { DialogService } from '~shared/dialog';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageProviders } from '~core/list-page/list-page-providers.class';
+
 
 @Component({
 	selector: 'find-products-dialog-app',
@@ -16,26 +17,28 @@ import { ListPageProviders } from '~core/list-page/list-page-providers.class';
 	styleUrls: ['./find-products-dialog.component.scss'],
 	// changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders('products-page', ERM.PRODUCT),
-		CommonDialogService,
-		{ provide: ERM_TOKEN, useValue: ERM.PRODUCT }
+		ListPageDataService,
+		ListPageViewService,
+		SelectionWithFavoriteService,
+		CommonDialogService
 	]
 })
 export class FindProductsDialogComponent extends TrackingComponent implements OnInit, AfterViewInit {
 
 	@Input() initialSelectedProducts: Product[];
-	@Input() submitCallback: Function;
+	@Input() submitCallback: (sel: { selectedProducts, unselectedProducts }) => Observable<any>;
 	searchFilterElements$: Observable<any[]>;
-
 	unselectedProducts: { [key: string]: Product } = {};
+
 	constructor(
-		protected router: Router,
-		protected cdr: ChangeDetectorRef,
-		protected featureSrv: ProductService,
-		protected viewSrv: ListPageViewService<Product>,
+		public router: Router,
+		public cdr: ChangeDetectorRef,
+		public featureSrv: ProductService,
+		public viewSrv: ListPageViewService<Product>,
 		public dataSrv: ListPageDataService<Product, ProductService>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService
+		public selectionSrv: SelectionWithFavoriteService,
+		public commonDlgSrv: CommonDialogService,
+		public dlgSrv: DialogService
 	) {
 		super();
 	}
@@ -47,6 +50,7 @@ export class FindProductsDialogComponent extends TrackingComponent implements On
 			initialSortBy: 'category.name'
 		});
 		this.dataSrv.init();
+		this.viewSrv.setup(ERM.PRODUCT);
 	}
 
 	ngAfterViewInit() {
@@ -56,7 +60,7 @@ export class FindProductsDialogComponent extends TrackingComponent implements On
 	}
 
 	getSelectedProducts() {
-		return Array.from(this.selectionSrv.selection.values());
+		return this.selectionSrv.getSelectionValues();
 	}
 
 	hasSelectedProducts() {
@@ -75,16 +79,14 @@ export class FindProductsDialogComponent extends TrackingComponent implements On
 	}
 
 	closeDlg() {
-		this.commonDlgSrv.close();
+		this.dlgSrv.close();
 	}
 
 	submit() {
 		// we add each project one by one to the store
-		const selectedProducts = this.getSelectedProducts();
-		const unselectedProducts = Object.keys(this.unselectedProducts).map(key => this.unselectedProducts[key]);
+		const selectedProducts = this.selectionSrv.getSelectionValues();
+		const unselectedProducts = Object.values(this.unselectedProducts);
 		this.submitCallback({ selectedProducts, unselectedProducts })
-			.subscribe(() => {
-				this.commonDlgSrv.close();
-			});
+			.subscribe(() => this.dlgSrv.close());
 	}
 }
