@@ -1,27 +1,23 @@
-import { NgModuleRef, OnInit } from '@angular/core';
+import { OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { SampleService, UserService } from '~entity-services';
-import { ERM, Sample } from '~models';
-import { DialogService } from '~shared/dialog';
-import { SearchService } from '~shared/filters';
-import { ListPageComponent } from '~core/list-page/list-page.component';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
-import { ListPageProviders } from '~core/list-page/list-page-providers.class';
+import { SampleService, UserService } from '~entity-services';
+import { ERM, Sample } from '~models';
 import { TrackingComponent } from '~utils/tracking-component';
 
 /** since we use the sample component on different pages, this page will keep the methods clean */
 export abstract class AbstractSampleCommonComponent extends TrackingComponent implements OnInit {
 	constructor(
-		protected router: Router,
-		protected userSrv: UserService,
-		protected featureSrv: SampleService,
-		protected viewSrv: ListPageViewService<Sample>,
+		public router: Router,
+		public userSrv: UserService,
+		public featureSrv: SampleService,
+		public viewSrv: ListPageViewService<Sample>,
 		public dataSrv: ListPageDataService<Sample, SampleService>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService
+		public selectionSrv: SelectionWithFavoriteService,
+		public commonDlgSrv: CommonDialogService
 	) {
 		super();
 	}
