@@ -1,39 +1,26 @@
-import { Router } from '@angular/router';
-import { OnInit, AfterViewInit } from '@angular/core';
+import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
+import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { GlobalServiceInterface } from '~entity-services/_global/global.service';
 import { EntityMetadata } from '~models';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageDataService } from '~core/list-page/list-page-data.service';
-import { ListPageViewService } from '~core/list-page/list-page-view.service';
-import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
-import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
+import { SelectionService } from '~core/list-page';
 
-export abstract class AbstractDataManagementComponent<T extends { id?: string },
-	G extends GlobalServiceInterface<T>> extends TrackingComponent implements OnInit {
+export abstract class AbstractDataManagementComponent extends TrackingComponent {
 
 	constructor(
-		protected router: Router,
-		protected featureSrv: G,
-		protected viewSrv: ListPageViewService<T>,
-		public dataSrv: ListPageDataService<T, G>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService,
+		public selectionSrv: SelectionService,
+		public commonDlgSrv: CommonDialogService,
 		public entityMetadata: EntityMetadata
 	) {
 		super();
 	}
 
-	ngOnInit() {
-		this.dataSrv.setup({
-			featureSrv: this.featureSrv,
-			searchedFields: ['name'],
-			initialSortBy: 'name'
+	mergeSelected() {
+		const items = Array.from(this.selectionSrv.getSelectionIds());
+		this.commonDlgSrv.openMergeDialog({
+			type: this.entityMetadata,
+			entities: items
 		});
-		this.dataSrv.init();
 	}
 
-	mergeSelected() {
-		const items = Array.from(this.selectionSrv.selection.keys());
-		this.commonDlgSrv.openMergeDialog({ type: this.entityMetadata, entities: items });
-	}
 }
