@@ -1,13 +1,12 @@
-import { ChangeDetectionStrategy, Component, NgModuleRef } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { AbstractDataManagementComponent } from '~features/data-management/containers/abstract-data-management.component';
-import { TagManagememtService } from '~features/data-management/services/tag-management.service';
-import { ERM, Tag, ERM_TOKEN } from '~models';
-import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
+import { TagService } from '~core/entity-services';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
+import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
+import { AbstractDataManagementComponent } from '~features/data-management/containers/abstract-data-management.component';
+import { ERM, Tag } from '~models';
 
 @Component({
 	selector: 'tag-data-management-page-app',
@@ -15,21 +14,33 @@ import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-provid
 	styleUrls: ['./tag-data-management-page.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.TAG, ERM.TAG),
-		CommonDialogService,
-		{ provide: ERM_TOKEN, useValue: ERM.TAG }
+		ListPageDataService,
+		ListPageViewService,
+		SelectionWithFavoriteService,
+		CommonDialogService
 	]
 })
-export class TagDataManagementPageComponent extends AbstractDataManagementComponent<Tag, TagManagememtService> {
+export class TagDataManagementPageComponent extends AbstractDataManagementComponent
+	implements OnInit {
 
 	constructor(
-		protected router: Router,
-		protected featureSrv: TagManagememtService,
-		protected viewSrv: ListPageViewService<Tag>,
-		public dataSrv: ListPageDataService<Tag, TagManagememtService>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService
+		public router: Router,
+		public tagSrv: TagService,
+		public viewSrv: ListPageViewService<Tag>,
+		public dataSrv: ListPageDataService<Tag, TagService>,
+		public selectionSrv: SelectionWithFavoriteService,
+		public commonDlgSrv: CommonDialogService
 	) {
-		super(router, featureSrv, viewSrv, dataSrv, selectionSrv, commonDlgSrv, ERM.TAG);
+		super(selectionSrv, commonDlgSrv, ERM.TAG);
+	}
+
+	ngOnInit() {
+		this.dataSrv.setup({
+			featureSrv: this.tagSrv,
+			searchedFields: ['name'],
+			initialSortBy: 'name'
+		});
+		this.dataSrv.init();
+		this.viewSrv.setup(this.entityMetadata);
 	}
 }
