@@ -1,13 +1,12 @@
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
 import { AbstractTaskCommonComponent } from '~common/task';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
-import { ListPageProviders, ProviderKey } from '~core/list-page/list-page-providers.class';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
 import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { TaskService, UserService } from '~entity-services';
-import { ERM, Task } from '~models';
-import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
+import { Task } from '~models';
 
 @Component({
 	selector: 'product-tasks-app',
@@ -15,21 +14,23 @@ import { CommonDialogService } from '~common/dialog/services/common-dialog.servi
 	styleUrls: ['./product-tasks.component.scss'],
 	changeDetection: ChangeDetectionStrategy.OnPush,
 	providers: [
-		ListPageProviders.getProviders(ProviderKey.PRODUCT_TASK, ERM.TASK),
-	]
+		ListPageDataService,
+		ListPageViewService,
+		SelectionWithFavoriteService,
+		CommonDialogService]
 })
 
 export class ProductTasksComponent extends AbstractTaskCommonComponent
 	implements OnInit {
 	constructor(
-		protected route: ActivatedRoute,
-		protected userSrv: UserService,
-		protected router: Router,
-		protected featureSrv: TaskService,
-		protected viewSrv: ListPageViewService<Task>,
+		public route: ActivatedRoute,
+		public userSrv: UserService,
+		public router: Router,
+		public featureSrv: TaskService,
+		public viewSrv: ListPageViewService<Task>,
 		public dataSrv: ListPageDataService<Task, TaskService>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService
+		public selectionSrv: SelectionWithFavoriteService,
+		public commonDlgSrv: CommonDialogService
 	) {
 		super(
 			router,
