@@ -1,36 +1,35 @@
-import { Component, NgModuleRef, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { ProjectFeatureService } from '~features/project/services/project-feature.service';
-import { Project, ERM_TOKEN, ERM } from '~models';
 import { CommonDialogService } from '~common/dialog/services/common-dialog.service';
-import { FilterType } from '~shared/filters';
-import { ThumbService } from '~shared/rating/services/thumbs.service';
-import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
 import { ListPageDataService } from '~core/list-page/list-page-data.service';
 import { ListPageViewService } from '~core/list-page/list-page-view.service';
+import { SelectionWithFavoriteService } from '~core/list-page/selection-with-favorite.service';
+import { ProjectFeatureService } from '~features/project/services/project-feature.service';
+import { ERM, Project } from '~models';
+import { FilterType } from '~shared/filters';
 import { TrackingComponent } from '~utils/tracking-component';
-import { ListPageProviders } from '~core/list-page/list-page-providers.class';
 
 @Component({
 	selector: 'projects-page-app',
 	templateUrl: './projects-page.component.html',
 	styleUrls: ['./projects-page.component.scss'],
 	providers: [
-		ListPageProviders.getProviders('projects-page', ERM.PROJECT),
-		CommonDialogService,
-		{ provide: ERM_TOKEN, useValue: ERM.PROJECT }]
+		ListPageDataService,
+		ListPageViewService,
+		SelectionWithFavoriteService,
+		CommonDialogService
+	]
 })
 export class ProjectsPageComponent extends TrackingComponent implements OnInit {
 	filterTypes = [FilterType.CREATED_BY];
 
 	constructor(
-		protected router: Router,
-		protected featureSrv: ProjectFeatureService,
-		protected viewSrv: ListPageViewService<Project>,
+		public router: Router,
+		public featureSrv: ProjectFeatureService,
+		public viewSrv: ListPageViewService<Project>,
 		public dataSrv: ListPageDataService<Project, ProjectFeatureService>,
-		protected selectionSrv: SelectionWithFavoriteService,
-		protected commonDlgSrv: CommonDialogService,
-		protected thumbSrv: ThumbService
+		public selectionSrv: SelectionWithFavoriteService,
+		public commonDlgSrv: CommonDialogService,
 	) {
 		super();
 	}
@@ -42,5 +41,6 @@ export class ProjectsPageComponent extends TrackingComponent implements OnInit {
 			initialSortBy: 'name'
 		});
 		this.dataSrv.init();
+		this.viewSrv.setup(ERM.PROJECT);
 	}
 }
